Ensure util.get_architecture() runs only once
util.get_architecture() recently was wrapped using python3's lru_cache()
which will cache the result so we only invoke 'dpkg --print-architecture'
once.  In practice, cloud-init.log will show multiple invocations of the
command.  The source of this was that the debian Distro object implements
the get_primary_arch() with this command, but it was not calling it from
util, but issuing a util.subp() directly.  This branch also updates
cc_apt_configure methods to fetch the arch value from the distro class,
and then ensure that the methods apt_configure calls pass the arch value
around.

- Fix get_architecture to not cache the value if target is not /.
  dpkg does report the correct target arch (ie x86_64 host, i386
  in a 32-bit chroot).  This matches lsb_release/_lsb_release.