Ensure util.get_architecture() runs only once (#172)
* Ensure util.get_architecture() runs only once

util.get_architecture() recently was wrapped using python3's lru_cache()
which will cache the result so we only invoke 'dpkg --print-architecture'
once.  In practice, cloud-init.log will show multiple invocations of the
command.  The source of this was that the debian Distro object implements
the get_primary_arch() with this command, but it was not calling it from
util, but issuing a util.subp() directly.  This branch also updates
cc_apt_configure methods to fetch the arch value from the distro class,
and then ensure that the methods apt_configure calls pass the arch value
around.

* utils: remove lsb_release and get_architecture wrappers

The original lsb_release wrapper was used to prevent polluting the
single value we cached, however lru_cache() already handles this
case by using args, kwargs values to cache different calls to the
method.

* rename_apt_list: use all positional parameters