tools/read-version: don't enforce version parity in release branch CI
We have a bootstrapping problem with new releases, currently.  To take
the example of 20.1: the branch that bumps the version fails CI because
there is no 20.1 tag for it to use in read-version.  Previously, this
was solved by creating a tag and pushing it to the cloud-init repo
before the commit landed.  However, we have GitHub branch protection
enabled, so the commit that needs to be tagged is not created until the
pull request lands in master.

This works around this problem by introducing a very specific check: if
we are performing CI for an upstream release branch, we skip the
read-version checking that we know will fail.