cc_growpart: refactor ResizeGrowPart to parse output
Now that we're logging growpart output, utilize this data to report the
before, after size values of the device after resizing.   Add a get_size
method to the Resize objects, Gpart retains use of the original get_size
method which opens and seeks to end of the file.  For growpart we drop
the --dry-run, and instead capture the stdout, stderr of the result and
parse the before and after size values from there.

Add a unittest to verify parsing and error handling.