Update tooling for GitHub-based new releases (#223)
* tools/read-version: don't enforce version parity in release branch CI

We have a bootstrapping problem with new releases, currently.  To take
the example of 20.1: the branch that bumps the version fails CI because
there is no 20.1 tag for it to use in read-version.  Previously, this
was solved by creating a tag and pushing it to the cloud-init repo
before the commit landed.  However, we have GitHub branch protection
enabled, so the commit that needs to be tagged is not created until the
pull request lands in master.

This works around this problem by introducing a very specific check: if
we are performing CI for an upstream release branch, we skip the
read-version checking that we know will fail.

* tools/make-tarball: add --version parameter

When using make-tarball as part of a CI build of a new upstream release,
the version it determines is inconsistent with the version that other
tools determine.  Instead of encoding the logic here (as well as in
Python elsewhere), we add a parameter to allow us to set it from outside
the script.

* packages/bddeb: handle missing version_long in new version CI

If we're running in CI for a new upstream release, we have to use
`version` instead of `version_long` (because we don't yet have the tag
required to generate `version_long`).