Fix docs for OpenStack DMI Asset Tag (#228)
In cloud-init 19.2, we added the ability for cloud-init to detect
OpenStack platforms by checking for "OpenStack Compute" or "OpenStack
Nova" in the chassis asset tag.  However, this was never reflected
in the documentation.  This patch updates the datasources documentation
for OpenStack to reflect the possibility of using the chassis asset tag.

LP: #1669875