Fallback to cached datasource if no valid datasource
If no valid datasource was found after the search for
network datasources, we assume our datasource has gone
away and try to use the last cached datasource if
it hasn't been removed from the configuration.

An example where this is useful is when the EC2
IMDS endpoint gets disabled after the 1st boot.