Prefer shorting NIC names when sorting.
When cloud-init guesses the first NIC to bring it up and retrieve
network meta-data, it used alphabetic sorting (corrected for numbers to
ensure that 10 is larger than 9).
On some clouds, secondary and later NICs are attached to a secondary
PCI bus (behind a PCI brige) -- on Ubuntu, we'll then see ensX for the
first NIC and enpYsZ for secondary NICs. Unfortunately enp < ens, so the
sorting did not yield the wanted result.

This approach simplifies the sorting. We first sort by length of the NIC
name (shorter names come first) and then alphabetically.
So: ens3 < ens9 < eth0 < eth8 < ens10 < eth10 < enp4s4 < enp4s9 < enp5s2
No guarantee to always guess correctly this way, but better.
The priority on length maintains then correct handling of numbers (9<10).

Signed-off-by: Kurt Garloff <scs@garloff.de>