Only consider letters when counting length.
This is proposal 2, so we don't change enp0s10 vs enp1s9 sorting
compared to the previous behavior.
We still prefer shorter names, but only consider alphabetic letters and
leave the original logic to split numbers from letters in place.
Not as simple as the previous version, but introduces fewer change.

Signed-off-by: Kurt Garloff <scs@garloff.de>