Fix cloud-init ignoring some mimetypes in user-data.
This patch if from Ryan Harper (inspired by my bug report and a
suggestion from Scott Moser) from the discussion on PR #234.

cloud-init would not accept unexpected mime-types in multipart user-data
for the cloud-config piece before; now it does. Problem was observed
and reproduced with old heat on OTC and is now fixed.

Signed-off-by: Kurt Garloff <kurt@garloff.de>