Fix cloud-init ignoring some misdeclared mimetypes in user-data.
On some platforms (old heat on OpenTelekomCloud), the user-data mime part
is mislabeled x-shellscript.  cloud-init would not accept this unexpected
mime-type in multipart user-data.  Cloud-init will now run find_ctype()
on the content of the mime-part to check if it matches known include types.

This patch is from Ryan Harper (inspired by my bug report and a
suggestion from Scott Moser) from the discussion on PR #234.


Signed-off-by: Kurt Garloff <kurt@garloff.de>