distros: handle a potential mirror filtering error case
As written, it's possible that the first transformation for a mirror
hostname could be passed None if the parsed mirror URL didn't have a
hostname component, when the defined interface is that the
transformations will be passed strings.

This isn't an error currently, because the first transformation happens
to gracefully handle being passed None.  It returns None, so the
pipeline processing ends there.

This was caught when testing out mypy on the cloud-init codebase.