test_azure: rewrite helpers to use bare asserts
We're seeing a strange error in xenial builds which appear to be
stemming from the AssertionError that pytest produces being _different_
from the standard AssertionError.  This means that the helpers modified
here weren't behaving correctly, because they weren't catching
AssertionErrors as one would expect.

(I believe this is related, in some way, to
https://github.com/pytest-dev/pytest/issues/645, but the only version of
pytest where we're affected is so far in the past that it's not worth
pursuing it any further as we have a workaround.)