travis: cache the chroot we use for package builds
Prior to this change, we would debootstrap for every single integration
test run in Travis.  This changes that, so we will store the chroot
created by debootstrap in Travis' cache, and use it if available.  This
saves 2-3 minutes, or ~1/3rd of the integration test run time (which is
our longest run).