.travis.yml: only store new schroot if something has changed
Prior to this change, the process of tarring up would mean that Travis
would always detect that the cache had changed, and we would incur ~30s
of packing/transferring at the end of every build.

Instead, we now check if there was any change to the installed contents
of the schroot, and only generate a new tarball if there were changes.