.travis.yml: rationalise installation of dependencies in host (#449)
We build in a schroot using sbuild, so we don't need Build-Depends
installed in the host; save some time (and manual maintenance when
Build-Depends change) by dropping their installation and skipping the
check for them when calling `bddeb`.  We _do_ require debhelper to be
installed to build the source package, so ensure it is installed.

Additionally, while I was here, replace ubuntu-dev-tools with
devscripts; we don't need everything that ubuntu-dev-tools pulls in.