Moving swap tests from TestFstabHandling to TestSwapFileCreation
Creating a new class of tests TestSwapFileCreation() for this purpose.
The class still uses test_helpers.FilesystemMockingTestCase to mock
a file system so swap file can be created.

Also added a check on cloudinit/config/cc_mounts.py to make sure the
swap file really exists, which is not the case on unittests. The swap
file is not being created, the test just makes sure if the correct calls
are being done.

Signed-off-by: Eduardo Otubo <otubo@redhat.com>