test_handler_ca_certs.py: Testcase for deleting existing CAs and adding new one
Existing testcases do not handle the situation where all the existing CA
certs are deleted (so resulting in an empty /etc/ca-certificates.conf file
and when one, or more, new CAs are added). This testcase verifies that the
resultant ca-certificates.conf file does not contain any blank lines which
are known to cause problems for the update-ca-certificates utility.