cli: add devel make-mime subcommand (#518)
* cli: add devel make-mime subcommand

Cloud-init documents an in-source-tree tool, make-mime.py used to
help users create multi-part mime user-data.  This tool is not shipped
in the cloud-init install and unavailable at runtime.  This patch
takes tools/make-mime.py and makes the functionality available via
the devel subcommand.

The primary interface of --attach file:content-type is still present.
The cli now adds:

 -l, --list-types  Print out a list of supported content-types
 -f, --force       Ignore errors for unsupported content-types

The tool will now raise a RunTime error if the supplied content-type
is not supported (or more likely a typo:
  x-shell-script vs.  x-shellscript)

* make-mime: write to stderr and exit 1 instead of raising RuntimeError

* Update example to match docs

* Update docs for make-mime subcommand

* Remove tools/make-mime.py; replaced by cloud-init devel make-mime

Co-authored-by: Rick Harding <rharding@mitechie.com>