debian/cloud-init.postinst: fix NVME grub install device on upgrade (#538)
In [0] we landed a change to the determination of the grub install device
for new instance launches, specifically to fix the grub install device
determined for instances booting from NVMe drives.  However, for running
NVMe-root instances to successfully install grub updates which change
the ABI between the core and its modules, we also need to fix the grub
install device that the previous (incorrect) code determined.

This commit performs that fixing, on upgrade.

[0] https://github.com/canonical/cloud-init/commit/fc07d633f7cb694423349a2c4b10c91c4b4981a2

LP: #1889555

Co-authored-by: Daniel Watkins <oddbloke@ubuntu.com>