bddeb: add --packaging-branch argument
bddeb normally builds a .deb package using the template packaging files
in packages/debian/.

The new --packaging-branch argument allows to specify a git branch where
to pull the packaging (i.e. the debian/ directory) from. This is useful
to build a .deb package from master with the very same packaging which
is used for the uploads. The contents of debian/patches is excluded.