Do not reset to dhcp network when DS is None
As per below bug in launchpad
We use config drive(/dev/sr0) as a data source to configure network
interfaces in the guest but configdrive is not always available and may be
removed after couple of hours from the hypervisor.

On a first boot cloudinit uses data provided in config drive and updates system
level network scripts /etc/sysconfig/ifcfg-* (Static configuration of networks)
files and also configures the interface in the guest.

As long as the configdrive is available, reboots will relay on system scripts
for the configuring network but once configdrive is removed,
datasource becomes None meaning it neither system script nor config drive
which makes cloud init to configure default network which is DHCP

https://bugs.launchpad.net/cloud-init/+bug/1893770sdd