Correct documentation and testcases for the following user-data YAML:
For cc_users_groups the user setting "expiredate" needs to be a string to
which is then passed to the useradd command.

For cc_users_groups, the user setting "inactive" does not actually disable
accounts, the useradd "--inactive" option actually defines the number of
days after password expiry that users can still login. So I have changed
the docs to show it taking a quoted value of days (which works with the
current code) rather than a boolean value.

freebsd.py, which doesn't use useradd, *already* has broken handling of
"inactive" as it passes a non-existant "-E" option to the "pw useradd"
command - it should actually pass "-e" instead followed by "+<DAYS>d".
I talked to @goneri and @OddBloke about this on IRC.

The existing testcases for users and groups haven't picked up on
"expiredate" and "inactive" not working as they don't check/mock
/etc/shadow where these values end up. I see the testcases using
"getent passwd" but "getent shadow" will not work when running testcases
as a non-root user.