split read_fs_info into linux & freebsd parts
FreeBSD doesn't have blkid, so we want to use geom to list devices and
their fstypes and labels.

This PR also adds `jail` to the list of is_container()
And we now also properly cache geom and blkid output!

A test is added to verify the new behaviour by correctly identifying
NoCloud on FreeBSD.