Update documentation for 'groups' directive
Hi,

I'm not sure whether this is actually a bug, bug the 'groups' sub-property of 'user' does not add the user to multiple groups. It only adds the user to the first group in that list in addition to its primary username == groupname  group. If you want to add a user to multiple groups, you must use the top-level 'groups' directive.

Thx,
Matthias