Fix stacktrace in DataSourceRbxCloud if no metadata disk is found.
Largely speaking, ds-identify protects from this scenario being
hit, but if DataSourceRbxCloud ran and there was no metadata
disks found (LABEL=CLOUDMD), then it would stacktrace.

The fix is just to clean up the get_md function a little
bit, and the explicitly check for False as a return value in _get_data.