ds-identify: Call blkid with a list of devices.
On a system with many block devices (or "virtual" block devices such as
lvm/device-mapper ... ), running blkid can be very slow.  ds-identify is
using it to identify filesystems that may have a config disk of different
types.

If we accept limiting the search to "real" block devices based on name,
then we dramatically reduce ds-identify's load in early boot when lvm and
other devices are coming up.

This limits the search by name to those devices that do not match:

    dm-* md*

For reference, running blkid on non-existant block device still
exits success if it found any devices.
This will fail:

  $ blkid -c /dev/null -o export /dev/does-not-exist ; echo $?
  2

But this will succeed as long as any device in the list existed:

  $ blkid -c /dev/null -o export /dev/sd* /dev/does-not-exist ; echo $?
  0