only run a subset of integration tests in CI
This introduces the "ci" mark, used to indicate a test which should run
as part of our CI integration testing run and the `integration-tests-ci`
tox environment, which runs only those tests.  Travis has been adjusted
to use this tox environment.

(All current module tests have been marked with the "ci" mark, but the
one bug test that we have has not.)