integration_tests: restrict test_lxd_bridge appropriately
On xenial, the bridge test fails because xenial's LXD doesn't include
the `network` subcommand.  On bionic, the bridge test fails within
containers, because LXD isn't able to manipulate the host kernel as it
expects.

(focal and later do run successfully in containers, but we don't have a
good way of expressing that presently.)