doc: move testing section from HACKING to its own doc (#739)
Our HACKING doc is very long, making it hard to justify expanding or
restructuring its content too much. This moves the testing section to
its own doc, and adds section headers in to make finding/linking to
testing guidelines easier.

To avoid confusion, the title of debugging.rst is changed.