Fix the TestApt tests using apt-key on Xenial and Hirsute
* Xenial issue

The `apt-key finger` format changed since Xenial. Sample Xenial output:

pub   4096R/991BC93C 2018-09-17
      Key fingerprint = F6EC B376 2474 EDA9 D21B  7022 8719 20D1 991B C93C

Sample Focal output:

pub   rsa4096 2016-04-12 [SC]
      EB4C 1BFD 4F04 2F6D DDCC  EC91 7721 F63B D38B 4796

What didn't change is the format of the key fingerprint, which should be
enough to ensure that the right key is in place across all the supported
releases.

* Hirsute issue

TestApt::test_ppa_source also fails on Hirsute because of a difference
in how the PPA keys are added. On Focla this command:

  add-apt-repository ppa:simplestreams-dev/trunk

install /etc/apt/trusted.gpg.d/simplestreams-dev_ubuntu_trunk.gpg, while
on Hirsute the file is names simplestreams-dev-ubuntu-trunk.gpg. The
filename is part of the `apt-key finger` output, and this the test
fails. Only checking for the presence of the key fingerprint in apt-key
also covers this case.