Refactor release handing in integration tests
There are currently two ways we filter the releases supported by a given
test:
1. Mark the test with a pytest mark for the given release.
2. Check the release at the beginning of the test at skip if appropriate
(1) is problematic because it's quite rigid. There's no easy way to
specify things like "all versions greater than focal", or "only run on
Xenial and Groovy" without adding yet another mark. Having dozens of
marks for each combination of release we may want isn't realistic.
(2) is problematic because by the time we hit the first line of the
test, we've already wasted time launching an instance.

This change allows us to use a single decorator that can take a defined
release or list of defined releases.